+++ packages/partitura_core/lib/partitura_core.dart
@@ -61,6 +61,7 @@ export 'src/theory/pitch.dart';
 export 'src/theory/scale.dart';
 export 'src/theory/set_theory.dart';
 export 'src/theory/time_signature.dart';
+export 'src/theory/tone_row.dart';
 export 'src/theory/transposition.dart';
 export 'src/theory/triad.dart';
 export 'src/theory/tuning.dart';